apps/digital-twin/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
databricks-sql-connector==3.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sse-starlette==1.8.2
python-multipart==0.0.6
